+++ app/layout.tsx
@@ -1,6 +1,9 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import { ClerkProvider } from "@clerk/nextjs";
+
+import { Header } from "@/components/Header";
+
 import "./globals.css";
 
 const geistSans = Geist({
@@ -29,7 +32,10 @@ export default function RootLayout({
         lang="en"
         className={`${geistSans.variable} ${geistMono.variable} h-full antialiased`}
       >
-        <body className="min-h-full flex flex-col">{children}</body>
+        <body className="min-h-full flex flex-col">
+          <Header />
+          <div className="flex flex-1 flex-col">{children}</div>
+        </body>
       </html>
     </ClerkProvider>
   );